converted to .cpp/.hpp
