+++ buoy_gazebo/src/PolytropicPneumaticSpring/CMakeLists.txt
@@ -1,4 +1,4 @@
 gz_add_plugin(PolytropicPneumaticSpring
   SOURCES
-    PolytropicPneumaticSpring.cc
+    PolytropicPneumaticSpring.cpp
 )
